feat: suggested changes implemented
